+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 - Improved IP detection  (thanks to [Ludovic Riaudel](https://github.com/lriaudel)).
 - [WP-CLI] `wp apcu settings` and `wp apcu analytics` now support `--stdout` flag.
+- [WP-CLI] Improved `wp apcu analytics` yaml formatting.
 - [WP-CLI] Improved documentation.
 
 ### Fixed